feat: Add model execution tracking phase 1

+++ backend/app/scoring_engine/base_score/winner_score.py
@@ -1,12 +1,12 @@
-WINNER_POINTS_CORRECT = 5
-WINNER_POINTS_INCORRECT = 0
+WINNER_POINTS_CORRECT = 5.0
+WINNER_POINTS_INCORRECT = 0.0
 
 
 def calculate_winner_score(
     prediction: dict,
     actual_result: dict,
     config: dict | None = None,
-) -> int:
+) -> float:
     predicted_winner = prediction["match_prediction"]["predicted_winner"]
     actual_winner = actual_result["actual_winner"]
 
@@ -14,5 +14,5 @@ def calculate_winner_score(
     points_incorrect = config.get("winner_points_incorrect", WINNER_POINTS_INCORRECT) if config else WINNER_POINTS_INCORRECT
 
     if predicted_winner == actual_winner:
-        return points_correct
-    return points_incorrect
+        return float(points_correct)
+    return float(points_incorrect)
